Assert no futures left for current fiber only
We keep information about each created future and when fiber exits, we
check if there are futures left behind. When more than one fiber is
created, we keep info for each fiber's futures in the same object, so each
fiber thinks it has N futures, while in fact some of them are for the
other fiber. This leads to a wrong assertion error. Use new array to keep
information about each fiber's futures and assert only the futures from current fiber are finished.